Fix bugs in merge
